More fixes to config install

+++ php/install.sh
@@ -2,7 +2,7 @@
 
 . ../config.sh
 
-if ../tools/iamservertype -q php-cgi; then
+if iamdaemon php5-fpm; then
 
 	# Note: the order of --exclude & --include matters here... (we only want non-hidden .php files)
 	rsync --recursive --delete --times --exclude=".*" --include="*/" --include="*.php" --exclude="*" ./ $WWW_DIR/php/;
